+++ action_server/tests/action_server_tests/test_server_headers_in_body/my_package/action.py
@@ -1,0 +1,11 @@
+from sema4ai.actions import Request, action
+from sema4ai.actions._secret import Secret
+
+
+@action
+def my_action(arg: str, private_info: Secret, request: Request) -> str:
+    """
+    Does something
+    """
+    assert "x-action-context" in request.headers
+    return "Ok"